refactor(LottoTicket): 로또 티켓 다형성 적용
